+++ R/write_colony.R
@@ -200,23 +200,10 @@ write_colony <- function(
 
   # Filename -------------------------------------------------------------------
   # Get date and time to have unique filenaming
+  file.date <- format(Sys.time(), "%Y%m%d@%H%M")
+
   if (is.null(filename)) {
-    filename <- stringi::stri_join(
-      "radiator_colony_",
-      stringi::stri_replace_all_fixed(
-        Sys.time(),
-        pattern = " EDT",
-        replacement = "",
-        vectorize_all = FALSE
-      ) %>%
-        stringi::stri_replace_all_fixed(
-          str = .,
-          pattern = c("-", " ", ":"),
-          replacement = c("", "@", ""),
-          vectorize_all = FALSE
-        ) %>%
-        stringi::stri_sub(str = ., from = 1, to = 13)
-    )
+    filename <- stringi::stri_join("radiator_colony_", file.date)
 
     if (!is.null(imputation.method)) {
       filename.imp <- stringi::stri_replace_all_fixed(