+++ NEWS.md
@@ -1,3 +1,7 @@
+# radiator v.0.0.8 2019-02-08
+
+* New output file: fineRADstructure
+
 # radiator v.0.0.7 2019-02-06
 
 * Better VCF parsing